+++ client/src/views/home.js
@@ -1,6 +1,5 @@
 import { div, ul, li, a, span, button, small, p, strong } from '@cycle/dom'
 import { yaml, ago, showDesc, pluralize } from './util'
-import ordinal from 'ordinal'
 
 const perPage = 10
 
@@ -28,7 +27,7 @@ const home = ({ feed, feedStart, feedActive, unitf, obalance, cbalance, channels
   ])])
 
 const itemRenderer = ({ feedActive, unitf, expert }) => ([ type, ts, msat, obj ]) => {
-  const fid     = `${type}-${obj.id || obj.pay_index}`
+  const fid     = `${type}-${obj.payment_hash}`
       , visible = fid == feedActive
       , tsStr   = new Date(ts*1000).toLocaleString()
 
@@ -42,10 +41,8 @@ const itemRenderer = ({ feedActive, unitf, expert }) => ([ type, ts, msat, obj ]
       li([ strong(type == 'in' ? 'Received:' : 'Sent:'), ' ', tsStr ])
     , type == 'in' && obj.msatoshi_received > obj.msatoshi ? li([ strong('Overpayment:'), ' ', unitf(obj.msatoshi_received-obj.msatoshi) ]) : ''
     , type == 'out' && obj.msatoshi ? li([ strong('Fee:'), ' ', feesText(obj, unitf) ]) : ''
-    , type == 'out' && obj.route ? li([ strong('Route:'), ' ', obj.route.length > 1 ? `${obj.route.length} hops` : 'direct payment'
-                                                        , ' ', small(`(${ordinal(obj.sendpay_tries)} attempt)`) ]) : ''
     , showDesc(obj) ? li([ strong('Description:'), ' ', span('.break-word', obj.description) ]) : ''
-    , type == 'out' ? li([ strong('Destination:'), ' ', small('.break-all', obj.destination) ]) : ''
+    , type == 'out' && obj.destination ? li([ strong('Destination:'), ' ', small('.break-all', obj.destination) ]) : ''
     , li([ strong('Payment hash:'), ' ', small('.break-all', obj.payment_hash) ])
     , expert ? li(yaml(obj)) : ''
     ])