+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Support multi-part payments, grouped together in the UI (#147)
+
 - Fix bug in automatic credentials generation (#145)
 
 ## 0.2.14 - 2020-04-22